OF-2428: Allow MUC occupants to request each-other's VCards
This implements a hack that redirects IQ requests from one MUC occupant to another to the bare JID of the occupant, rather than the full JID. This allows the home server of the occupant (rather than its client) to process the request. This is in-line with how XEP-0054 defines how VCards should be requested.