OF-2444: Use dedicated lock for org.jivesoftware.openfire.http.HttpSession#sentElements
Access to org.jivesoftware.openfire.http.HttpSession#sentElements was previously guarded by a lock that was re-used for other purposes.

By replacing the lock for org.jivesoftware.openfire.http.HttpSession#sentElements with one that is dedicted the access of that field, the probability of resource contention related to the lock is reduced.