OF-2444: Asynchronously deliver outbound data over BOSH
When sending stanzas back to the client, this is now done in an asynchronous manner, outside of a lock held on the connectionQueue. The delivery of inbound data already occurred in an asynchronous manner. The threadpool that was used for that is now being used for both.

To achieve this, the processing of (outbound) stanzas has been decoupled from the management of connectionQueue (adding/removing connections), allowing it to happen outside of the guard of a lock.