OF-2479: Allow clients that do websockets without the required XMPP framing
This basically is a hack to get Tsung to connect to Openfire of websockets.

The functionality introduced here is switched off by default. If enabled, it replaces `stream` element names with `open` and `close` where appropriate, to conform to 'XMPP framing' as described in https://datatracker.ietf.org/doc/html/rfc7395#section-3.3