OF-2423: Improve removal of client sessions from cache
Prior to this, the failure to identify a client session as a (regular) user session resulted in the session being assumed an anonymous user session. When a session is removed twice (which should not occur, but does occur with websockets for some reason), this causes a second client route to be removed from the cache.

This commit ensures that a session is indeed an anonymous session before assuming that it is.

For websockets, a warning will still be logged, as there are still two attempts to remove the session. Optimizing this is the subject of another issue: OF-2485